Clean up LoggingManager script
